Use an API route to search domains
This allows us to cut cut client-side dependencies
