+++ app/components/SearchBar.tsx
@@ -3,15 +3,15 @@
 import { useCluster } from '@providers/cluster';
 import { useTokenRegistry } from '@providers/token-registry';
 import { TokenInfoMap } from '@solana/spl-token-registry';
-import { Connection } from '@solana/web3.js';
 import { Cluster } from '@utils/cluster';
-import { getDomainInfo, hasDomainSyntax } from '@utils/name-service';
 import { LOADER_IDS, LoaderName, PROGRAM_INFO_BY_ID, SPECIAL_IDS, SYSVAR_IDS } from '@utils/tx';
 import bs58 from 'bs58';
 import { useRouter, useSearchParams } from 'next/navigation';
 import React, { useId } from 'react';
 import { Search } from 'react-feather';
 import Select, { ActionMeta, InputActionMeta, ValueType } from 'react-select';
+
+import { FetchedDomainInfo } from '../api/domain-info/[domain]/route';
 
 interface SearchOptions {
     label: string;
@@ -22,6 +22,10 @@ interface SearchOptions {
     }[];
 }
 
+const hasDomainSyntax = (value: string) => {
+    return value.length > 4 && value.substring(value.length - 4) === '.sol';
+};
+
 export function SearchBar() {
     const [search, setSearch] = React.useState('');
     const searchRef = React.useRef('');
@@ -31,7 +35,7 @@ export function SearchBar() {
     const selectRef = React.useRef<Select<any> | null>(null);
     const router = useRouter();
     const { tokenRegistry } = useTokenRegistry();
-    const { url, cluster, clusterInfo } = useCluster();
+    const { cluster, clusterInfo } = useCluster();
     const searchParams = useSearchParams();
     const onChange = ({ pathname }: ValueType<any, false>, meta: ActionMeta<any>) => {
         if (meta.action === 'select-option') {
@@ -58,7 +62,7 @@ export function SearchBar() {
         setSearchOptions(options);
 
         // checking for non local search output
-        if (hasDomainSyntax(search)) {
+        if (hasDomainSyntax(search) && cluster === Cluster.MainnetBeta) {
             // if search input is a potential domain we continue the loading state
             domainSearch(options);
         } else {
@@ -72,9 +76,8 @@ export function SearchBar() {
     // appends domain lookup results to the local search state
     const domainSearch = async (options: SearchOptions[]) => {
         setLoadingSearchMessage('Looking up domain...');
-        const connection = new Connection(url);
         const searchTerm = search;
-        const updatedOptions = await buildDomainOptions(connection, search, options);
+        const updatedOptions = await buildDomainOptions(search, options);
         if (searchRef.current === searchTerm) {
             setSearchOptions(updatedOptions);
             // after attempting to fetch the domain name we can conclude the loading state
@@ -213,8 +216,9 @@ function buildTokenOptions(search: string, cluster: Cluster, tokenRegistry: Toke
     }
 }
 
-async function buildDomainOptions(connection: Connection, search: string, options: SearchOptions[]) {
-    const domainInfo = await getDomainInfo(search, connection);
+async function buildDomainOptions(search: string, options: SearchOptions[]) {
+    const domainInfoResponse = await fetch(`/api/domain-info/${search}`);
+    const domainInfo = await domainInfoResponse.json() as FetchedDomainInfo;
     const updatedOptions: SearchOptions[] = [...options];
     if (domainInfo && domainInfo.owner && domainInfo.address) {
         updatedOptions.push({
