+++ app/components/account/DomainsCard.tsx
@@ -3,8 +3,10 @@
 import { Address } from '@components/common/Address';
 import { ErrorCard } from '@components/common/ErrorCard';
 import { LoadingCard } from '@components/common/LoadingCard';
-import { DomainInfo, useUserDomains } from '@utils/name-service';
+import { useUserDomains } from '@utils/name-service';
 import React from 'react';
+
+import { DomainInfo } from '@/app/utils/domain-info';
 
 export function DomainsCard({ address }: { address: string }) {
     const [domains, domainsLoading] = useUserDomains(address);
