+++ app/utils/name-service.tsx
@@ -2,9 +2,6 @@
 
 import {
     getFilteredProgramAccounts,
-    getHashedName,
-    getNameAccountKey,
-    getNameOwner,
     NAME_PROGRAM_ID,
     performReverseLookup,
 } from '@bonfida/spl-name-service';
@@ -13,42 +10,7 @@ import { Connection, PublicKey } from '@solana/web3.js';
 import { Cluster } from '@utils/cluster';
 import { useEffect, useState } from 'react';
 
-// Address of the SOL TLD
-const SOL_TLD_AUTHORITY = new PublicKey('58PwtjSDuFHuUkYjH9BYnnQKHfwo9reZhC2zMJv9JPkx');
-
-export interface DomainInfo {
-    name: string;
-    address: PublicKey;
-}
-export const hasDomainSyntax = (value: string) => {
-    return value.length > 4 && value.substring(value.length - 4) === '.sol';
-};
-
-async function getDomainKey(name: string, nameClass?: PublicKey, nameParent?: PublicKey) {
-    const hashedDomainName = await getHashedName(name);
-    const nameKey = await getNameAccountKey(hashedDomainName, nameClass, nameParent);
-    return nameKey;
-}
-
-// returns non empty wallet string if a given .sol domain is owned by a wallet
-export async function getDomainInfo(domain: string, connection: Connection) {
-    const domainKey = await getDomainKey(
-        domain.slice(0, -4), // remove .sol
-        undefined,
-        SOL_TLD_AUTHORITY
-    );
-    try {
-        const registry = await getNameOwner(connection, domainKey);
-        return registry && registry.registry.owner
-            ? {
-                  address: domainKey.toString(),
-                  owner: registry.registry.owner.toString(),
-              }
-            : null;
-    } catch {
-        return null;
-    }
-}
+import { DomainInfo, SOL_TLD_AUTHORITY } from './domain-info';
 
 async function getUserDomainAddresses(connection: Connection, userAddress: string): Promise<PublicKey[]> {
     const filters = [
